requirements.txt
et-xmlfile==1.1.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
